refactor(router): rename lib/router.tsx to lib/navigation.tsx
Avoids the name collision with pages/router.tsx (the router instance).
The lib module holds navigation helpers and route-state hooks, not a
router; the split itself stays — lib/navigation.tsx is deliberately
route-tree-free so anything can import it without pulling in the page
graph.
