+++ apps/webapp/src/modules/convert/components/PsmConversionWidgetPane.tsx
@@ -4,7 +4,7 @@ import { useConfigContext } from '@/modules/config/hooks/useConfigContext';
 import { QueryParams } from '@/lib/constants';
 import { ConvertIntent } from '@/lib/enums';
 import { useNavigate } from '@tanstack/react-router';
-import { keepSearch, useAppSearchParams, useRouteConvertIntent, useRouteIntent } from '@/lib/router';
+import { keepSearch, useAppSearchParams, useRouteConvertIntent, useRouteIntent } from '@/lib/navigation';
 import { Intent } from '@/lib/enums';
 import { useMemo } from 'react';
