+++ apps/webapp/src/modules/stusds/components/StUSDSWidgetPane.tsx
@@ -4,7 +4,7 @@ import { QueryParams, REFRESH_DELAY } from '@/lib/constants';
 import { SharedProps } from '@/modules/app/types/Widgets';
 import { useConfigContext } from '@/modules/config/hooks/useConfigContext';
 import { useNavigate } from '@tanstack/react-router';
-import { keepSearch, useAppSearchParams, useRouteExpertIntent } from '@/lib/router';
+import { keepSearch, useAppSearchParams, useRouteExpertIntent } from '@/lib/navigation';
 import { useSubgraphUrl } from '@/modules/app/hooks/useSubgraphUrl';
 import { ExpertIntent } from '@/lib/enums';
