+++ apps/webapp/src/widgets/shared/components/ui/card/InteractiveStatsCardWithVaultAccordion.tsx
@@ -11,7 +11,7 @@ import { HStack } from '@/widgets/shared/components/ui/layout/HStack';
 import { ArrowRight } from 'lucide-react';
 import { formatUnits } from 'viem';
 import { formatBigInt, formatNumber } from '@/utils';
-import { AppLink } from '@/lib/router';
+import { AppLink } from '@/lib/navigation';
 import { InteractiveStatsCard } from './InteractiveStatsCard';
 import { PriceData } from '@/hooks';
